feat: expose version comparison from core
