+++ .changeset/tidy-version-comparison.md
@@ -1,0 +1,6 @@
+---
+"@stll/folio-core": minor
+"@stll/folio-agents": patch
+---
+
+Expose document version comparison from folio-core while preserving the agents API.
